Revert "fix(buffer): check for errors using errors.Is() incase of wrapping"
This reverts commit 5cb9db476b1a7d877f2ab1bcb1b0a976f3f0fcb2.

The right logic is to ignore if the error is NewFile, indicating that a
new file was created. NewFile is sort of like io.EOF error, in the sense
that it depends upon the caller to determine if it's indeed an error or
not.

The mentioned commit errors out if it is indeed a NewFile, which
introduces a bug.

Moreover, this statement should not be checking for wrapped errors.  It
is possible that a new file error was wrapped by another error. In that
case, we should indeed error out.