Add a test to ensure that calling Set after Close does not panic
On a previous version of ristretto, we experienced a panic occasionally due to calling Set after Close resulting in a write to the closed setBuf channel. We originally wrote this test as part of a fix on a forked version of Ristretto. Thankfully, on HEAD, this test is no longer failing, so we are submitting it to act as a guard against any regressions in the future.